feat: update websocket integrations
