+++ api/socket.js
@@ -3,8 +3,8 @@ import AWSLink from "./awslink";
 
 // Set the variables
 // yeah, not gonna bother cleaning the code
-const { awslink, doctortoken } = AWSLink();
-const endpoint = "ws://https://" + awslink + "/doctor/chat";
+const { doctortoken } = AWSLink();
+const endpoint = "wss://capstone-project.duckdns.org:8080/doctor/chat";
 
 // Define the socket
 const socket = io(endpoint, {
