+++ api/websocket.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Websocket() {
+  return (
+    <div>websocket</div>
+  )
+}
