--- app/chat-test/socketio/page.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function Page() {
-  return (
-    <div>Page</div>
-  )
-}
