+++ app/dashboard-dokter/chat/page.jsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState } from "react";
+import { useState, useRef, useEffect } from "react";
 import Image from "next/image";
 import {
    createReceipt,
@@ -10,6 +10,7 @@ import Link from "next/link";
 import { useAtom } from "jotai";
 import ChatSidebar from "./chatSidebar";
 import Chat from "./chat";
+import Action from "./action";
 
 export default function Page() {
    const [selectedFile, setSelectedFile] = useState(null);
@@ -39,63 +40,53 @@ export default function Page() {
 
    const [users, setUsers] = useState([
       {
-         id: "0",
+         id: "1",
          nama: "rudi",
          chat: [
-            { message: "Hola", type: "received" },
-            { message: "Hola", type: "sent" },
-            { message: "?", type: "received" },
-            { message: "ko", type: "sent" },
-         ],
-      },
-      {
-         id: "1",
-         nama: "rudix",
-         chat: [
-            { message: "tes", type: "received" },
-            { message: "tes2", type: "received" },
-            { message: "tes", type: "sent" },
-            { message: "tes2", type: "sent" },
+            { message: "Hola", type: "user" },
+            { message: "Hola", type: "doctor" },
+            { message: "?", type: "user" },
+            { message: "ko", type: "doctor" },
          ],
       },
    ]);
    const handleUserClick = (id) => {
       setSelectedUser(id);
    };
-   const handleSendMessage = (e) => {
-      e.preventDefault();
+   //    const handleSendMessage = (e) => {
+   //       e.preventDefault();
 
-      const updatedUsers = users.map((user) => {
-         if (newMessage === "" && selectedFile === "") {
-            return user;
-         } else {
-            if (user.id === selectedUser) {
-               console.log(selectedFile);
-               if (selectedFile !== null) {
-                  return {
-                     ...user,
-                     chat: [
-                        ...user.chat,
-                        { message: selectedFile, type: "sent" },
-                     ],
-                  };
-               } else {
-                  return {
-                     ...user,
-                     chat: [
-                        ...user.chat,
-                        { message: newMessage, type: "sent" },
-                     ],
-                  };
-               }
-            }
-         }
-         return user;
-      });
+   //       const updatedUsers = users.map((user) => {
+   //          if (newMessage === "" && selectedFile === "") {
+   //             return user;
+   //          } else {
+   //             if (user.id === selectedUser) {
+   //                console.log(selectedFile);
+   //                if (selectedFile !== null) {
+   //                   return {
+   //                      ...user,
+   //                      chat: [
+   //                         ...user.chat,
+   //                         { message: selectedFile, type: "sent" },
+   //                      ],
+   //                   };
+   //                } else {
+   //                   return {
+   //                      ...user,
+   //                      chat: [
+   //                         ...user.chat,
+   //                         { message: newMessage, type: "sent" },
+   //                      ],
+   //                   };
+   //                }
+   //             }
+   //          }
+   //          return user;
+   //       });
 
-      setUsers(updatedUsers);
-      setNewMessage("");
-   };
+   //       setUsers(updatedUsers);
+   //       setNewMessage("");
+   //    };
 
    const handleSendObat = async (e) => {
       e.preventDefault();
@@ -122,6 +113,63 @@ export default function Page() {
       setNewMessage("");
    };
 
+   // Websocket Starts here ---- Cleaning the code above will take a lot of time
+
+   const [chatHistory, setChatHistory] = useState([]);
+
+   const socket = useRef(null);
+
+   useEffect(() => {
+      const createSocket = () => {
+         const socket = new WebSocket("ws://localhost:8080");
+
+         // Event listener for WebSocket connection establishment
+         socket.addEventListener("open", () => {
+            console.log("WebSocket connection established");
+         });
+
+         // Event listener for receiving messages from the server
+         socket.addEventListener("message", (event) => {
+            const receivedMessage = JSON.parse(event.data);
+            setChatHistory((prevChatHistory) => [
+               ...prevChatHistory,
+               receivedMessage,
+            ]);
+         });
+
+         // Event listener for WebSocket disconnections
+         socket.addEventListener("close", () => {
+            console.log("WebSocket connection closed");
+         });
+
+         return socket;
+      };
+
+      // Only create a new socket if the previous one is not defined
+      if (!socket.current) {
+         socket.current = createSocket();
+      }
+
+      return () => {
+         // Close the WebSocket connection when the component is unmounted
+         if (socket.current) {
+            socket.current.close();
+         }
+      };
+   }, [socket]);
+
+   const handleSendMessage = (e) => {
+      e.preventDefault();
+      if (newMessage.trim() === "") return;
+
+      // Send the message to the WebSocket server
+      const messageToSend = {
+         "to": 1,
+         "message": newMessage,
+         "type": "doctor",
+      };
+      socket.current.send(JSON.stringify(messageToSend));
+   };
    return (
       <>
          {receiptLocal.length !== 0 ? (
@@ -182,62 +230,11 @@ export default function Page() {
                         </svg>
                         <div className="text-neutral-0 text-xs">Action</div>
                      </button>
-                     {isDropdownOpen && (
-                        <ul className="absolute z-10 right-0 bg-web-green-300 top-12  w-[150px] rounded-3xl">
-                           <li>
-                              <Link
-                                 href="#"
-                                 className=" py-3 px-3 flex items-center gap-5 cursor-pointer"
-                              >
-                                 <svg
-                                    width="18"
-                                    height="18"
-                                    viewBox="0 0 18 18"
-                                    fill="none"
-                                    xmlns="http://www.w3.org/2000/svg"
-                                 >
-                                    <path
-                                       d="M6.625 6.125L9.12492 8.62492M9.12492 8.62492L11.6249 11.1249M9.12492 8.62492L6.625 11.1249M9.12492 8.62492L11.6249 6.125M9.125 16.125C4.98287 16.125 1.625 12.7672 1.625 8.625C1.625 4.48287 4.98287 1.125 9.125 1.125C13.2672 1.125 16.625 4.48287 16.625 8.625C16.625 12.7672 13.2672 16.125 9.125 16.125Z"
-                                       stroke="white"
-                                       strokeWidth="2"
-                                       strokeLinecap="round"
-                                       strokeLinejoin="round"
-                                    />
-                                 </svg>
-                                 <p className="text-neutral-0 text-xs">
-                                    Tutup sesi
-                                 </p>
-                              </Link>
-                           </li>
-
-                           <li>
-                              <Link
-                                 href="/dashboard-dokter/obat-dokter"
-                                 className=" py-3 px-3 flex items-center gap-5 cursor-pointer"
-                              >
-                                 <svg
-                                    width="18"
-                                    height="19"
-                                    viewBox="0 0 18 19"
-                                    fill="none"
-                                    xmlns="http://www.w3.org/2000/svg"
-                                 >
-                                    <path
-                                       d="M17.4514 10.1615C16.8154 8.83853 15.7423 7.8363 14.4306 7.27506C15.5833 5.55123 15.3845 3.22606 13.8741 1.74276C13.0394 0.94098 11.9663 0.5 10.8533 0.5H10.8136C9.66091 0.5 8.58773 0.981069 7.75304 1.78285L1.23447 8.51782C0.439526 9.35969 -0.0374428 10.4421 0.00230455 11.6047C0.00230455 12.7673 0.479273 13.8497 1.27422 14.6915C2.10891 15.5334 3.22184 15.9343 4.29502 15.9343C5.24896 15.9343 6.16315 15.6136 6.95809 15.0122C6.99784 15.0523 6.99784 15.1325 7.03759 15.1726C7.99152 17.2171 10.0186 18.5 12.2445 18.5C13.1189 18.5 13.9139 18.2996 14.7088 17.9388C17.5706 16.5356 18.8028 13.088 17.4514 10.1615ZM15.5038 10.1615L8.23001 13.6091C8.03127 12.8474 8.07102 12.0056 8.34925 11.2439C8.70698 10.2016 9.46218 9.35969 10.4559 8.87862C11.0123 8.598 11.6085 8.47773 12.2047 8.47773C13.5164 8.51782 14.7486 9.15924 15.5038 10.1615ZM8.94546 2.94543C9.46218 2.42428 10.0981 2.14365 10.8136 2.14365C11.529 2.14365 12.2047 2.42428 12.7215 2.90534C13.7549 3.94766 13.7946 5.6314 12.7612 6.71381L12.642 6.83408C12.5227 6.83408 12.3637 6.83408 12.2445 6.83408C11.37 6.83408 10.5751 7.03452 9.78016 7.39532C9.38268 7.59577 8.98521 7.8363 8.66723 8.11693L6.28239 5.75167L8.94546 2.94543ZM6.20289 13.4488C5.16946 14.4911 3.50007 14.5312 2.42689 13.4889C1.91018 12.9677 1.63195 12.3263 1.63195 11.6047C1.63195 10.8831 1.91018 10.2016 2.38715 9.6804L5.08997 6.91425L7.51456 9.3196C7.23632 9.72049 6.99784 10.2016 6.7991 10.6826C6.52087 11.4844 6.40163 12.2862 6.48112 13.1281L6.20289 13.4488ZM13.9934 16.4154C13.4369 16.696 12.8407 16.8163 12.2445 16.8163C10.9328 16.8163 9.74041 16.1748 8.94546 15.1726L16.2192 11.7249C16.6564 13.569 15.782 15.5334 13.9934 16.4154Z"
-                                       fill="white"
-                                    />
-                                 </svg>
-                                 <p className="text-neutral-0 text-xs">
-                                    Resep obat
-                                 </p>
-                              </Link>
-                           </li>
-                        </ul>
-                     )}
+                     <Action isDropdownOpen={isDropdownOpen} />
                   </div>
                </div>
                <div className="h-[70vh] overflow-y-auto">
-                <Chat users={users} selectedUser={selectedUser}/>
+                  <Chat message={chatHistory} />
                </div>
 
                <div className="flex-grow px-2">
